+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.15.0] - 2018-03-01
+
+### Changed
+
+- Variant data moved into component config
+  - For details see the [variant docs](./docs/variant.md)
+
 ## [0.14.0] - 2018-02-10
 
 ### Changed
@@ -12,7 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - More flexible configuration lookup
   - The config file is now retrieved via [cosmiconfig](https://github.com/davidtheclark/cosmiconfig)
   - The default config file is `uiengine.config.js`
-  - FOr alternative ways to store your config (i.e. in package.json) see the cosmiconfig docs
+  - For alternative ways to store your config (i.e. in package.json) see the cosmiconfig docs
 
 ## [0.13.0] - 2018-02-09